Cleaned up the toolset package.
